core: extract audit append helpers
